+++ src/main/java/step2/domain/frame/Frame.java
@@ -1,5 +1,6 @@
-package step2.domain;
+package step2.domain.frame;
 
+import step2.domain.Pitch;
 import step2.domain.state.State;
 
 public interface Frame {
