+++ src/main/java/step2/domain/frame/Frames.java
@@ -1,4 +1,4 @@
-package step2.domain;
+package step2.domain.frame;
 
 import java.util.ArrayList;
 import java.util.List;
